Organize imports & fix Eclipse warnings

+++ gerrit-server/src/test/java/com/google/gerrit/server/mail/receive/data/NonUTF8Message.java
@@ -46,7 +46,7 @@ public class NonUTF8Message extends RawMailMessage {
     int[] arr = new int[raw.length()];
     int i = 0;
     for (char c : raw.toCharArray()) {
-      arr[i++] = (int) c;
+      arr[i++] = c;
     }
     return arr;
   }
